% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_auc_fast}
\alias{calc_auc_fast}
\title{fast auc calculation using sorted approach}
\usage{
calc_auc_fast(expression, truth)
}
\arguments{
\item{expression}{vector of expression values}

\item{truth}{vector of ground truth (0/1 or logical)}
}
\value{
auc value
}
\description{
fast auc calculation using sorted approach
}
