% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_bulk_auc_one_sample_fast}
\alias{calc_bulk_auc_one_sample_fast}
\title{fast bulk auc calculation using sorted approach}
\usage{
calc_bulk_auc_one_sample_fast(
  bulk_vector,
  sample_name,
  training_matrix,
  training_genes,
  threshold_list = NULL,
  sep = "r"
)
}
\arguments{
\item{bulk_vector}{a vector of gene expression in the sample (with named entries)}

\item{sample_name}{a character vector, the name of a sample being processed}

\item{training_matrix}{a genes x cell-types ground truth matrix}

\item{training_genes}{a vector of gene names to calculate the auc with}

\item{threshold_list}{not used in sorted approach (for compatibility)}

\item{sep}{a character value to split the cell type from the replicate number}
}
\value{
roc-auc value for this bulk sample
}
\description{
fast bulk auc calculation using sorted approach
}
