% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fdr}
\alias{get_fdr}
\title{Calculates FDR when given a prediction matrix and a ground truth matrix}
\usage{
get_fdr(expression, truth, threshold)
}
\arguments{
\item{expression}{genes x cell-types gene expression matrix}

\item{truth}{genes x cell-types ground truth matrix}

\item{threshold}{a value to set the prediction for the expression matrix}
}
\value{
an FDR value for the full matrix
}
\description{
Calculates FDR when given a prediction matrix and a ground truth matrix
}
