% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fpr}
\alias{get_fpr}
\title{Calculates FPR when given a prediction matrix and a ground truth matrix}
\usage{
get_fpr(expression, truth, threshold)
}
\arguments{
\item{expression}{genes x cell-types gene expression matrix}

\item{truth}{genes x cell-types ground truth matrix}

\item{threshold}{a value to set the prediction for the expression matrix}
}
\value{
an FPR value for the full matrix
}
\description{
Calculates FPR when given a prediction matrix and a ground truth matrix
}
