% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fpr_vectorized}
\alias{get_fpr_vectorized}
\title{vectorized fpr calculation for multiple thresholds}
\usage{
get_fpr_vectorized(expression, truth, thresholds)
}
\arguments{
\item{expression}{vector of expression values}

\item{truth}{vector of ground truth (0/1 or logical)}

\item{thresholds}{vector of threshold values}
}
\value{
vector of fpr values for each threshold
}
\description{
vectorized fpr calculation for multiple thresholds
}
