% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_weights.R
\name{max_spm}
\alias{max_spm}
\title{Calculates specificity scores for a gene (from a genes x cell-type matrix) using SPM method
as described in Xiao SJ, Zhang C, Zou Q, Ji ZL. TiSGeD: a database for tissue-specific genes. Bioinformatics. 2010 May 1;26(9):1273-5. doi: 10.1093/bioinformatics/btq109. Epub 2010 Mar 11. PMID: 20223836; PMCID: PMC2859128.}
\usage{
max_spm(data)
}
\arguments{
\item{data}{a vector of gene expression values across cell-types}
}
\value{
a specificity score value (double)
}
\description{
Calculates specificity scores for a gene (from a genes x cell-type matrix) using SPM method
as described in Xiao SJ, Zhang C, Zou Q, Ji ZL. TiSGeD: a database for tissue-specific genes. Bioinformatics. 2010 May 1;26(9):1273-5. doi: 10.1093/bioinformatics/btq109. Epub 2010 Mar 11. PMID: 20223836; PMCID: PMC2859128.
}
