% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_weights.R
\name{Spm}
\alias{Spm}
\title{Calculates specificity scores for a gene (from a genes x cell-type matrix) using SPM method
as described in Xiao SJ, Zhang C, Zou Q, Ji ZL. TiSGeD: a database for tissue-specific genes. Bioinformatics. 2010 May 1;26(9):1273-5. doi: 10.1093/bioinformatics/btq109. Epub 2010 Mar 11. PMID: 20223836; PMCID: PMC2859128.
code adapted from Kryuchkova-Mostacci N, Robinson-Rechavi M. A benchmark of gene expression tissue-specificity metrics. Brief Bioinform. 2017 Mar 1;18(2):205-214. doi: 10.1093/bib/bbw008. PMID: 26891983; PMCID: PMC5444245.}
\usage{
Spm(data)
}
\arguments{
\item{data}{a vector of gene expression values across cell-types}
}
\value{
a specificity score value (double)
}
\description{
Calculates specificity scores for a gene (from a genes x cell-type matrix) using SPM method
as described in Xiao SJ, Zhang C, Zou Q, Ji ZL. TiSGeD: a database for tissue-specific genes. Bioinformatics. 2010 May 1;26(9):1273-5. doi: 10.1093/bioinformatics/btq109. Epub 2010 Mar 11. PMID: 20223836; PMCID: PMC2859128.
code adapted from Kryuchkova-Mostacci N, Robinson-Rechavi M. A benchmark of gene expression tissue-specificity metrics. Brief Bioinform. 2017 Mar 1;18(2):205-214. doi: 10.1093/bib/bbw008. PMID: 26891983; PMCID: PMC5444245.
}
