% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtraction.R
\name{subtract_single_log}
\alias{subtract_single_log}
\title{log transforms bulk and reference profiles, and the subtracts contaminating profiles}
\usage{
subtract_single_log(
  learning_rate,
  bulk,
  pseudobulk,
  cells_in_use,
  proportions_vector,
  anti_identity_vector,
  specificity_score
)
}
\arguments{
\item{learning_rate}{a value used to scale the subtraction magnitude}

\item{bulk}{a vector of gene expression in the sample (with named entries)}

\item{pseudobulk}{a genes x cell-types matrix of gene expression values from the single cell reference}

\item{cells_in_use}{a character vector of cell types being used for proportion modeling and subtraction}

\item{proportions_vector}{a vector of estimated proportions for the bulk sample (with entries named for the cell-types)}

\item{anti_identity_vector}{a vector of ones and zeros that matches the proportions vector in naming and length, with a zero value that corresponds to the target cell type (so that the target cell type is not subtracted)}

\item{specificity_score}{a vector of gene level weights calculated using the specificity score from the single cell reference}
}
\value{
a vector of gene expression values, named by the genes
}
\description{
log transforms bulk and reference profiles, and the subtracts contaminating profiles
}
