% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtraction.R
\name{subtraction}
\alias{subtraction}
\title{Function to iteratively subtract across samples}
\usage{
subtraction(
  bulk,
  reference,
  sep = "r",
  cell_types_matrix,
  training_matrix,
  specificity_weights,
  learning_rate = c(0, seq(1/(2^seq(0, 10, 1)))),
  max_iterations = 100,
  verbose = T
)
}
\arguments{
\item{bulk}{a vector of gene expression in the sample (with named entries)}

\item{reference}{a genes x cell-types matrix of gene expression values from the single cell reference}

\item{sep}{a character value to split the cell type from the replicate number in the sample name (assumed structure is cell-sep-number)}

\item{cell_types_matrix}{a samples x cell-types matrix describing the target and contaminant cell types to estimate per sample}

\item{training_matrix}{a genes x cell-types ground truth matrix}

\item{specificity_weights}{a vector of gene level weights calculated using the specificity score from the single cell reference}

\item{learning_rate}{list of values to use as the learning rates: ex: 1/(2**seq(0,10,1)) returns a list of 1, 1/2, 1/4, 1/8, etc...}

\item{max_iterations}{maximum number of iterations to go through}

\item{verbose}{if true, then the function will return information about each iteration}
}
\value{
a matrix of cleaned bulk expression
}
\description{
Function to iteratively subtract across samples
}
