% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_auc_fast}
\alias{calc_auc_fast}
\title{fast auc calculation using vectorized threshold processing}
\usage{
calc_auc_fast(expression, truth, threshold_list = c(0, seq(0, 15, 0.1)))
}
\arguments{
\item{expression}{vector of expression values}

\item{truth}{vector of ground truth (0/1 or logical)}

\item{threshold_list}{vector of thresholds to use}
}
\value{
auc value
}
\description{
fast auc calculation using vectorized threshold processing
}
